int64 indexing: complete loop counter migration and LongIndexBuffer
Loop counter migrations (int -> long) for all element iteration:
- Default.Reduction.Mean/Std/Var: loop over result.size
- np.random.randint/uniform: loop over data.Count
- np.all/np.any: axis reduction loops
- np.eye/np.mgrid: dimension iteration
- np.repeat: repeats iteration
- NDArray.unique: element counting
- np.nanmean/nanstd/nanvar: element and counting loops

LongIndexBuffer implementation:
- Replaces List<long> for nonzero indices (avoids int.MaxValue capacity limit)
- Unmanaged memory allocation with exponential growth
- Used in Default.NonZero.cs and ILKernelGenerator.Masking.cs
- Properly disposes native memory

IL Kernel improvements:
- ILKernelGenerator.Reduction.Axis.Simd: remove Parallel.For, use sequential loop
- ILKernelGenerator.Reduction.Axis.VarStd: remove Parallel.For, use sequential loop
- IKernelProvider: updated interface signatures for long indexing
- ILKernelGenerator.Masking: use LongIndexBuffer for coordinate collection

Test additions:
- NonzeroInt64Tests: verify nonzero works with long indices
- MatMulInt64Tests: verify matmul with long dimensions
- ArgsortInt64Tests: verify argsort with long indices

All 3913 tests pass (9 failures from documented dead code np.isinf).

+++ src/NumSharp.Core/Backends/Default/Math/Reduction/Default.Reduction.Mean.cs
@@ -28,7 +28,7 @@ namespace NumSharp.Backends
                 if (shape[axis] == 0)
                 {
                     result = np.empty(new Shape(resultShape), outputType);
-                    for (int i = 0; i < result.size; i++) result.SetAtIndex(double.NaN, i);
+                    for (long i = 0; i < result.size; i++) result.SetAtIndex(double.NaN, i);
                 }
                 else result = np.empty(new Shape(resultShape), outputType);
                 if (keepdims)
